Revert "stay-management用に調整"
